GameStore1

See merge request luka.shengelia758/gamestore!2

+++ GameStore/Services/GamesService.cs
@@ -1,7 +1,8 @@
-﻿using GameStore.Data;
+﻿using GameStore.CustomExceptions;
+using GameStore.Data;
 using GameStore.Models;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -18,34 +19,86 @@ namespace GameStore.Services
 
         public async Task<List<GameModel>> GetAllGames()
         {
-            return await _context.Games.ToListAsync();
+            List<GameModel> games = await _context.Games.ToListAsync();
+            if (games.Count == 0)
+            {
+                throw new DoesNotExistException("Games Not Found");
+            }
+            return games;
         }
         public async Task<GameModel> GetGameById(int id)
         {
-            return await _context.Games.Where(games => games.GameId == id).FirstOrDefaultAsync();
+            GameModel game = await _context.Games.Where(games => games.GameId == id).FirstOrDefaultAsync();
+            if (game == null)
+            {
+                throw new DoesNotExistException("Game not found with that ID");
+            }
+            return game;
         }
         public async Task<GameModel> AddGame(GameModel newGame)
         {
+            if (await _context.Games.Where(g => g.GameName == newGame.GameName).AnyAsync())
+            {
+                throw new AlreadyExistException("This game already exists in GameStore");
+            }
             await _context.Games.AddAsync(newGame);
             await _context.SaveChangesAsync();
 
-            var newlyAddedGame = await _context.Games.Where(game => game.GameName == newGame.GameName && game.GamePublisher == newGame.GamePublisher).FirstOrDefaultAsync();
-            return newlyAddedGame;
+            return newGame;
         }
-        public async Task<GameModel> EditGame([FromBody] GameModel game)
+        public async Task<GameModel> EditGame(GameModel editedGame, int id)
         {
-            var editgame = await _context.Games.Where(games => games.GameId == game.GameId).FirstOrDefaultAsync();
-            _context.Update(editgame.GameName);
+            var gameToUpdate = await _context.Games.Where(game => game.GameId == id).FirstOrDefaultAsync();
+            if (gameToUpdate == null)
+            {
+                throw new DoesNotExistException("Can not find a game with that ID to edit");
+            }
+            gameToUpdate.GameName = editedGame.GameName;
+            gameToUpdate.GameDescription = editedGame.GameDescription;
+            gameToUpdate.GameDeveloper = editedGame.GameDeveloper;
+            gameToUpdate.GamePublisher = editedGame.GamePublisher;
+            gameToUpdate.GamePrice = editedGame.GamePrice;
+            gameToUpdate.GameReleaseDate = editedGame.GameReleaseDate;
+            gameToUpdate.GameImgUrl = editedGame.GameImgUrl;
             await _context.SaveChangesAsync();
 
-            return editgame;
+            return gameToUpdate;
         }
         public async Task<GameModel> DeleteGame(int id)
         {
-            var findGame = await _context.Games.Where(x => x.GameId == id).FirstOrDefaultAsync();
+            GameModel findGame = await _context.Games.Where(x => x.GameId == id).FirstOrDefaultAsync();
+            if (findGame == null)
+            {
+                throw new DoesNotExistException("Can not find a game to delete");
+            }
             _context.Games.Remove(findGame);
             await _context.SaveChangesAsync();
             return findGame;
+        }
+        public async Task<List<GameModel>> FilterGamesByGenresAndName(List<int> genres, string name)
+        {
+            var result = new List<GameModel>();
+            if (!string.IsNullOrEmpty(name))
+            {
+                result = name.Length < 3 ? result :
+                    await _context.Games.Where(game => game.GameName.Contains(name)).ToListAsync();
+            }
+            else if (genres.Count != 0 && string.IsNullOrEmpty(name))
+            {
+                var filterGenres = _context.GamesAndGenres.Where(g => genres.Contains(g.GenreId));
+                result = await _context.Games.Where(game => filterGenres.Any(gen => gen.GameId == game.GameId)).ToListAsync();
+            }
+            else
+            {
+                var filterGenres = _context.GamesAndGenres.Where(g => genres.Contains(g.GenreId));
+                result = name.Length >= 3 ? await _context.Games.Where(game => filterGenres.Any(gen => gen.GameId == game.GameId) && game.GameName.Contains(name)).ToListAsync()
+                    : await _context.Games.Where(game => filterGenres.Any(gen => gen.GameId == game.GameId)).ToListAsync();
+            }
+            if (result.Count == 0)
+            {
+                throw new DoesNotExistException();
+            }
+            return result;
         }
     }
 }
